November 22, 2018 v1.3.1 - Sync with local. Made "src" dir. Use CMake Open Folder with VCPKG and VS2017.
(mingw/msys2 should still work but not tested fully yet)

+++ src/common_physics/opengl.cpp
@@ -15,7 +15,7 @@
 #include "utils.h"
 
 ///////////////////////////////////////////////////////////////////////////////
-void utSetOrthographicProjection(int scrW, int scrH)
+void utSetOrthographicProjection(float scrW, float scrH)
 {
 
     // switch to projection mode
